feat: add basic authentication to the put and delete routes

+++ tests/test_app.py
@@ -35,9 +35,10 @@ def test_list_users_with_users(client, user):
     assert response.json() == {'users': [user_schema]}
 
 
-def test_update_user(client, user):
+def test_update_user(client, user, token):
     response = client.put(
         '/users/1',
+        headers={'Authorization': f'Bearer {token}'},
         json={'username': 'update', 'email': 'update@example.com', 'password': 'psw'},
     )
 
@@ -48,27 +49,39 @@ def test_update_user(client, user):
     }
 
 
-def test_update_user_not_found(client):
+def test_update_user_not_found(client, token):
     response = client.put(
         '/users/100',
+        headers={'Authorization': f'Bearer {token}'},
         json={'username': 'update', 'email': 'update@example.com', 'password': 'psw'},
     )
 
-    assert response.status_code == HTTPStatus.NOT_FOUND
-    assert response.json() == {
-        'detail': 'User not found',
-    }
+    assert response.status_code == HTTPStatus.FORBIDDEN
 
 
-def test_delete_user(client, user):
-    response = client.delete('/users/1')
+def test_delete_user(client, user, token):
+    response = client.delete(
+        '/users/1',
+        headers={'Authorization': f'Bearer {token}'},
+    )
 
     assert response.status_code == HTTPStatus.OK
     assert response.json() == {'message': 'User deleted'}
 
 
-def test_delete_user_not_found(client):
-    response = client.delete('/users/100')
+def test_delete_user_not_found(client, token):
+    response = client.delete('/users/100', headers={'Authorization': f'Bearer {token}'})
 
-    assert response.status_code == HTTPStatus.NOT_FOUND
-    assert response.json() == {'detail': 'User not found'}
+    assert response.status_code == HTTPStatus.FORBIDDEN
+
+
+def test_get_token(client, user):
+    response = client.post(
+        '/token',
+        data={'username': user.email, 'password': user.clean_password},
+    )
+    token = response.json()
+
+    assert response.status_code == HTTPStatus.OK
+    assert 'access_token' in token
+    assert 'token_type' in token
